fix: datasets list settings shared with proposal datasets list (#2247)
## Description
This PR implements sharing the datasets list configuration (columns
selected, order and width) between the main datasets list and the
datasets list belonging to a proposal.

## Motivation
This Pr is part of the effort to improve UX by using the same settings
across different location showing the same information.

## Changes:
Please provide a list of the changes implemented by this PR
* datasets list component
* proposal dataset list component
* new dataset list columns service

## Tests included
- [x] Included for each change/fix?
- [ ] Passing? (Merge will not be approved unless this is checked) 

## Documentation
- [ ] swagger documentation updated \[required\] - not needed
- [ ] official documentation updated \[nice-to-have\]

## Summary by Sourcery

Share dataset list column configuration and rendering logic between the
main datasets table and proposal datasets table using a new reusable
service.

New Features:
- Introduce a shared datasets list service to centralize dataset column
configuration and rendering behavior.
- Apply shared dataset column settings to the proposal datasets table so
it uses the same configuration as the main datasets list, including
localization support.

Enhancements:
- Refactor dataset table component to delegate column conversion and
instrument name resolution to the shared service.
- Make common formatting pipes injectable so they can be used within the
shared datasets list service.

Tests:
- Extend end-to-end tests to verify default, personalized, and restored
dataset column configurations are consistently applied on both main and
proposal datasets lists.

---------
