+++ cypress/e2e/datasets/datasets-list-personalization.cy.js
@@ -1,4 +1,7 @@
-import { testData, defaultDatasetsColumnsList, personalizedDatasetsColumnsList } from "../../fixtures/testData";
+import {
+  defaultDatasetsColumnsList,
+  personalizedDatasetsColumnsList,
+} from "../../fixtures/testData";
 
 describe("1000 - Datasets list personalization", () => {
   beforeEach(() => {
@@ -9,11 +12,9 @@ describe("1000 - Datasets list personalization", () => {
     cy.removeDatasets();
   });
 
-  it("1010 - should be able to list datasets with default columns", () => {
-    const username = Cypress.env("username");
-    const password = Cypress.env("password");
-
+  it("1010 - should be able to visit main datasets list with default columns", () => {
     cy.createDataset({ type: "raw" });
+    cy.createProposal();
 
     cy.visit("/datasets");
 
@@ -21,136 +22,142 @@ describe("1000 - Datasets list personalization", () => {
 
     cy.finishedLoading();
 
-    cy.get('mat-table')
-      .find('mat-header-row.header')
-      .find('mat-header-cell')
-      .not('.cdk-column-row-checkbox')
-      .not('.cdk-column-table-menu')
-      .find('.mat-sort-header-content')
+    cy.get(".dataset-table mat-table")
+      .find("mat-header-row.header")
+      .find("mat-header-cell")
+      .not(".cdk-column-row-checkbox")
+      .not(".cdk-column-table-menu")
+      .find(".mat-sort-header-content")
       .each(($el, index) => {
         cy.wrap($el)
-          .invoke('text')
-          .invoke('trim')
-          .should('eq', defaultDatasetsColumnsList[index]);
+          .invoke("text")
+          .invoke("trim")
+          .should("eq", defaultDatasetsColumnsList[index]);
+      });
+  });
+
+  it("1015 - should be able to visit proposals datasets list with default columns", () => {
+    cy.visit("/proposals/20170266");
+
+    cy.finishedLoading();
+    cy.get("proposal-detail .general-header").should("exist");
+
+    cy.get(".mat-mdc-tab").contains("Datasets").click();
+
+    cy.finishedLoading();
+    cy.get(".proposal-dataset-table mat-table mat-header-row").should("exist");
+
+    cy.get(".proposal-dataset-table mat-table")
+      .find("mat-header-row.header")
+      .find("mat-header-cell")
+      .not(".cdk-column-row-checkbox")
+      .not(".cdk-column-table-menu")
+      .find(".mat-sort-header-content")
+      .each(($el, index) => {
+        cy.wrap($el)
+          .invoke("text")
+          .invoke("trim")
+          .should("eq", defaultDatasetsColumnsList[index]);
       });
   });
 
   it("1020 - should be able to personalize the list datasets with selected columns", () => {
-    //const username = Cypress.env("username");
-    //const password = Cypress.env("password");
-
     cy.visit("/datasets");
 
     cy.get(".dataset-table mat-table mat-header-row").should("exist");
 
     cy.finishedLoading();
 
-    cy.contains('mat-icon', 'more_vert')
-      .closest('button')
-      .click()
+    cy.contains("mat-icon", "more_vert").closest("button").click();
 
-    cy.contains('button.mat-mdc-menu-item', 'Column setting')
-      .click()
+    cy.contains("button.mat-mdc-menu-item", "Column setting").click();
 
     // Open column settings menu (already done earlier)
     // Enable columns
-    cy.contains('label', 'Start Time')
-      .closest('mat-checkbox')
-      .find('input')
-      .check({ force: true })
+    cy.contains("label", "Start Time")
+      .closest("mat-checkbox")
+      .find("input")
+      .check({ force: true });
 
-    cy.contains('label', 'End Time')
-      .closest('mat-checkbox')
-      .find('input')
-      .check({ force: true })
+    cy.contains("label", "End Time")
+      .closest("mat-checkbox")
+      .find("input")
+      .check({ force: true });
 
     // Disable Image column
-    cy.contains('label', 'Image')
-      .closest('mat-checkbox')
-      .find('input')
-      .uncheck({ force: true })
+    cy.contains("label", "Image")
+      .closest("mat-checkbox")
+      .find("input")
+      .uncheck({ force: true });
 
     // Apply
-    cy.contains('mat-icon', 'done')
-      .closest('button')
-      .click()
+    cy.contains("mat-icon", "done").closest("button").click();
 
-    cy.get('mat-table')
-      .find('mat-header-row.header')
-      .find('mat-header-cell')
-      .not('.cdk-column-row-checkbox')
-      .not('.cdk-column-table-menu')
-      .find('.mat-sort-header-content')
+    cy.get("mat-table")
+      .find("mat-header-row.header")
+      .find("mat-header-cell")
+      .not(".cdk-column-row-checkbox")
+      .not(".cdk-column-table-menu")
+      .find(".mat-sort-header-content")
       .each(($el, index) => {
         cy.wrap($el)
-          .invoke('text')
-          .invoke('trim')
-          .should('eq', personalizedDatasetsColumnsList[index]);
+          .invoke("text")
+          .invoke("trim")
+          .should("eq", personalizedDatasetsColumnsList[index]);
       });
   });
 
   it("1030 - should be able to save the personalized columns settings", () => {
-    //const username = Cypress.env("username");
-    //const password = Cypress.env("password");
-
     cy.visit("/datasets");
 
     cy.get(".dataset-table mat-table mat-header-row").should("exist");
 
     cy.finishedLoading();
 
-    cy.contains('mat-icon', 'more_vert')
-      .closest('button')
-      .click()
+    cy.contains("mat-icon", "more_vert").closest("button").click();
 
-    cy.contains('button.mat-mdc-menu-item', 'Column setting')
-      .click()
+    cy.contains("button.mat-mdc-menu-item", "Column setting").click();
 
     // Open column settings menu (already done earlier)
     // Enable columns
-    cy.contains('label', 'Start Time')
-      .closest('mat-checkbox')
-      .find('input')
-      .check({ force: true })
+    cy.contains("label", "Start Time")
+      .closest("mat-checkbox")
+      .find("input")
+      .check({ force: true });
 
-    cy.contains('label', 'End Time')
-      .closest('mat-checkbox')
-      .find('input')
-      .check({ force: true })
+    cy.contains("label", "End Time")
+      .closest("mat-checkbox")
+      .find("input")
+      .check({ force: true });
 
     // Disable Image column
-    cy.contains('label', 'Image')
-      .closest('mat-checkbox')
-      .find('input')
-      .uncheck({ force: true })
+    cy.contains("label", "Image")
+      .closest("mat-checkbox")
+      .find("input")
+      .uncheck({ force: true });
 
     // Apply
-    cy.contains('mat-icon', 'done')
-      .closest('button')
-      .click()
+    cy.contains("mat-icon", "done").closest("button").click();
 
     cy.finishedLoading();
 
-    cy.contains('mat-icon', 'more_vert')
-      .closest('button')
-      .click()
+    cy.contains("mat-icon", "more_vert").closest("button").click();
 
-    cy.contains('button.mat-mdc-menu-item', 'Save table setting')
-      .click()
+    cy.contains("button.mat-mdc-menu-item", "Save table setting").click();
 
     cy.visit("/datasets");
 
-    cy.get('mat-table')
-      .find('mat-header-row.header')
-      .find('mat-header-cell')
-      .not('.cdk-column-row-checkbox')
-      .not('.cdk-column-table-menu')
-      .find('.mat-sort-header-content')
+    cy.get("mat-table")
+      .find("mat-header-row.header")
+      .find("mat-header-cell")
+      .not(".cdk-column-row-checkbox")
+      .not(".cdk-column-table-menu")
+      .find(".mat-sort-header-content")
       .each(($el, index) => {
         cy.wrap($el)
-          .invoke('text')
-          .invoke('trim')
-          .should('eq', personalizedDatasetsColumnsList[index]);
+          .invoke("text")
+          .invoke("trim")
+          .should("eq", personalizedDatasetsColumnsList[index]);
       });
   });
 
@@ -163,7 +170,7 @@ describe("1000 - Datasets list personalization", () => {
     cy.finishedLoading();
 
     // log user out
-  
+
     cy.get(".user-button").should("contain.text", username).click();
 
     cy.get("[data-cy=logout-button]").click();
@@ -194,17 +201,42 @@ describe("1000 - Datasets list personalization", () => {
 
     cy.finishedLoading();
 
-    cy.get('mat-table')
-      .find('mat-header-row.header')
-      .find('mat-header-cell')
-      .not('.cdk-column-row-checkbox')
-      .not('.cdk-column-table-menu')
-      .find('.mat-sort-header-content')
+    cy.get("mat-table")
+      .find("mat-header-row.header")
+      .find("mat-header-cell")
+      .not(".cdk-column-row-checkbox")
+      .not(".cdk-column-table-menu")
+      .find(".mat-sort-header-content")
       .each(($el, index) => {
         cy.wrap($el)
-          .invoke('text')
-          .invoke('trim')
-          .should('eq', personalizedDatasetsColumnsList[index]);
+          .invoke("text")
+          .invoke("trim")
+          .should("eq", personalizedDatasetsColumnsList[index]);
+      });
+  });
+
+  it("1045 - should be able to visit proposals datasets list with personalized columns", () => {
+    cy.visit("/proposals/20170266");
+
+    cy.finishedLoading();
+    cy.get("proposal-detail .general-header").should("exist");
+
+    cy.get(".mat-mdc-tab").contains("Datasets").click();
+
+    cy.finishedLoading();
+    cy.get(".proposal-dataset-table mat-table mat-header-row").should("exist");
+
+    cy.get(".proposal-dataset-table mat-table")
+      .find("mat-header-row.header")
+      .find("mat-header-cell")
+      .not(".cdk-column-row-checkbox")
+      .not(".cdk-column-table-menu")
+      .find(".mat-sort-header-content")
+      .each(($el, index) => {
+        cy.wrap($el)
+          .invoke("text")
+          .invoke("trim")
+          .should("eq", personalizedDatasetsColumnsList[index]);
       });
   });
 
@@ -218,26 +250,23 @@ describe("1000 - Datasets list personalization", () => {
 
     cy.finishedLoading();
 
-    cy.contains('mat-icon', 'more_vert')
-      .closest('button')
-      .click()
+    cy.contains("mat-icon", "more_vert").closest("button").click();
 
-    cy.contains('button.mat-mdc-menu-item', 'Default setting')
-      .click()
+    cy.contains("button.mat-mdc-menu-item", "Default setting").click();
 
     cy.finishedLoading();
 
-    cy.get('mat-table')
-      .find('mat-header-row.header')
-      .find('mat-header-cell')
-      .not('.cdk-column-row-checkbox')
-      .not('.cdk-column-table-menu')
-      .find('.mat-sort-header-content')
+    cy.get("mat-table")
+      .find("mat-header-row.header")
+      .find("mat-header-cell")
+      .not(".cdk-column-row-checkbox")
+      .not(".cdk-column-table-menu")
+      .find(".mat-sort-header-content")
       .each(($el, index) => {
         cy.wrap($el)
-          .invoke('text')
-          .invoke('trim')
-          .should('eq', defaultDatasetsColumnsList[index]);
+          .invoke("text")
+          .invoke("trim")
+          .should("eq", defaultDatasetsColumnsList[index]);
       });
   });
 
@@ -279,25 +308,46 @@ describe("1000 - Datasets list personalization", () => {
 
     cy.finishedLoading();
 
-    cy.contains('mat-icon', 'more_vert')
-      .closest('button')
-      .click()
+    cy.contains("mat-icon", "more_vert").closest("button").click();
 
-    cy.contains('button.mat-mdc-menu-item', 'Save table setting')
-      .click()
+    cy.contains("button.mat-mdc-menu-item", "Save table setting").click();
 
-    cy.get('mat-table')
-      .find('mat-header-row.header')
-      .find('mat-header-cell')
-      .not('.cdk-column-row-checkbox')
-      .not('.cdk-column-table-menu')
-      .find('.mat-sort-header-content')
+    cy.get("mat-table")
+      .find("mat-header-row.header")
+      .find("mat-header-cell")
+      .not(".cdk-column-row-checkbox")
+      .not(".cdk-column-table-menu")
+      .find(".mat-sort-header-content")
       .each(($el, index) => {
         cy.wrap($el)
-          .invoke('text')
-          .invoke('trim')
-          .should('eq', defaultDatasetsColumnsList[index]);
+          .invoke("text")
+          .invoke("trim")
+          .should("eq", defaultDatasetsColumnsList[index]);
       });
   });
 
+  it("1065 - should be able to visit proposals datasets list with default columns", () => {
+    cy.visit("/proposals/20170266");
+
+    cy.finishedLoading();
+    cy.get("proposal-detail .general-header").should("exist");
+
+    cy.get(".mat-mdc-tab").contains("Datasets").click();
+
+    cy.finishedLoading();
+    cy.get(".proposal-dataset-table mat-table mat-header-row").should("exist");
+
+    cy.get(".proposal-dataset-table mat-table")
+      .find("mat-header-row.header")
+      .find("mat-header-cell")
+      .not(".cdk-column-row-checkbox")
+      .not(".cdk-column-table-menu")
+      .find(".mat-sort-header-content")
+      .each(($el, index) => {
+        cy.wrap($el)
+          .invoke("text")
+          .invoke("trim")
+          .should("eq", defaultDatasetsColumnsList[index]);
+      });
+  });
 });
