+++ src/app/datasets/dataset-table/dataset-table.component.spec.ts
@@ -10,6 +10,7 @@ import {
   mockDataset,
   createMock,
   MockActivatedRoute,
+  MockDatasetsListService,
 } from "shared/MockStubs";
 import { NO_ERRORS_SCHEMA } from "@angular/core";
 import {
@@ -51,6 +52,7 @@ import { DatePipe } from "@angular/common";
 import { FileSizePipe } from "shared/pipes/filesize.pipe";
 import { TitleCasePipe } from "shared/pipes/title-case.pipe";
 import { TranslateService } from "@ngx-translate/core";
+import { DatasetsListService } from "shared/services/datasets-list.service";
 
 const getConfig = () => ({});
 
@@ -89,10 +91,11 @@ describe("DatasetTableComponent", () => {
           ],
         }),
         { provide: TranslateService, useValue: { instant: (k: string) => k } },
-        JsonHeadPipe,
-        DatePipe,
+        //DatasetsListService,
+        // JsonHeadPipe,
+        // DatePipe,
         FileSizePipe,
-        TitleCasePipe,
+        // TitleCasePipe,
       ],
       declarations: [DatasetTableComponent],
     });
@@ -105,6 +108,7 @@ describe("DatasetTableComponent", () => {
           },
           { provide: DatasetsService, useClass: MockDatasetApi },
           { provide: ActivatedRoute, useClass: MockActivatedRoute },
+          { provide: DatasetsListService, useClass: MockDatasetsListService },
         ],
       },
     });
@@ -141,194 +145,6 @@ describe("DatasetTableComponent", () => {
 
       expect(emitSpy).toHaveBeenCalledTimes(1);
       expect(emitSpy).toHaveBeenCalledWith(dataset);
-    });
-  });
-
-  describe("#wipCondition()", () => {
-    xit("should...", () => {});
-  });
-
-  describe("#systemErrorCondition()", () => {
-    xit("should...", () => {});
-  });
-
-  describe("#userErrorCondition()", () => {
-    it("should return true if dataset has missingFilesError", () => {
-      const dataset = createMock<DatasetClass>({});
-      dataset.datasetlifecycle = {
-        archiveStatusMessage: "missingFilesError",
-      };
-
-      const userError = component.userErrorCondition(dataset);
-
-      expect(userError).toEqual(true);
-    });
-
-    it("should return false if dataset has no missingFilesError", () => {
-      const dataset = createMock<DatasetClass>({});
-      dataset.datasetlifecycle = {
-        archiveStatusMessage: "",
-      };
-
-      const userError = component.userErrorCondition(dataset);
-
-      expect(userError).toEqual(false);
-    });
-  });
-
-  describe("#archivableCondition()", () => {
-    it("should return false if dataset is not archivable and retrievable and does not have a missingFilesError", () => {
-      const dataset = createMock<DatasetClass>({});
-      dataset.datasetlifecycle = {
-        archivable: false,
-        retrievable: true,
-        archiveStatusMessage: "",
-      };
-
-      const archivable = component.archivableCondition(dataset);
-
-      expect(archivable).toEqual(false);
-    });
-
-    it("should return false if dataset is not archivable and retrievable and does have a missingFilesError", () => {
-      const dataset = createMock<DatasetClass>({});
-      dataset.datasetlifecycle = {
-        archivable: false,
-        retrievable: true,
-        archiveStatusMessage: "missingFilesError",
-      };
-
-      const archivable = component.archivableCondition(dataset);
-
-      expect(archivable).toEqual(false);
-    });
-
-    it("should return false if dataset is not archivable and not retrievable and does not have a missingFilesError", () => {
-      const dataset = createMock<DatasetClass>({});
-      dataset.datasetlifecycle = {
-        archivable: false,
-        retrievable: false,
-        archiveStatusMessage: "",
-      };
-
-      const archivable = component.archivableCondition(dataset);
-
-      expect(archivable).toEqual(false);
-    });
-
-    it("should return false if dataset is not archivable and not retrievable and does have a missingFilesError", () => {
-      const dataset = createMock<DatasetClass>({});
-      dataset.datasetlifecycle = {
-        archivable: false,
-        retrievable: false,
-        archiveStatusMessage: "missingFilesError",
-      };
-
-      const archivable = component.archivableCondition(dataset);
-
-      expect(archivable).toEqual(false);
-    });
-
-    it("should return false if dataset is not archivable and retrievable and does not have a missingFilesError", () => {
-      const dataset = createMock<DatasetClass>({});
-      dataset.datasetlifecycle = {
-        archivable: false,
-        retrievable: true,
-        archiveStatusMessage: "",
-      };
-
-      const archivable = component.archivableCondition(dataset);
-
-      expect(archivable).toEqual(false);
-    });
-
-    it("should return false if dataset is archivable and retrievable and does not have a missingFilesError", () => {
-      const dataset = createMock<DatasetClass>({});
-      dataset.datasetlifecycle = {
-        archivable: true,
-        retrievable: true,
-        archiveStatusMessage: "",
-      };
-
-      const archivable = component.archivableCondition(dataset);
-
-      expect(archivable).toEqual(false);
-    });
-
-    it("should return false if dataset is archivable and retrievable and does have a missingFilesError", () => {
-      const dataset = createMock<DatasetClass>({});
-      dataset.datasetlifecycle = {
-        archivable: true,
-        retrievable: true,
-        archiveStatusMessage: "missingFilesError",
-      };
-
-      const archivable = component.archivableCondition(dataset);
-
-      expect(archivable).toEqual(false);
-    });
-
-    it("should return true if dataset is archivable and not retrievable and does not have a missingFilesError", () => {
-      const dataset = createMock<DatasetClass>({});
-      dataset.datasetlifecycle = {
-        archivable: true,
-        retrievable: false,
-        archiveStatusMessage: "",
-      };
-
-      const archivable = component.archivableCondition(dataset);
-
-      expect(archivable).toEqual(true);
-    });
-  });
-
-  describe("#retrievableCondition()", () => {
-    it("should return false if dataset is archivable and not retrievable", () => {
-      const dataset = createMock<DatasetClass>({});
-      dataset.datasetlifecycle = {
-        archivable: true,
-        retrievable: false,
-      };
-
-      const retrievable = component.retrievableCondition(dataset);
-
-      expect(retrievable).toEqual(false);
-    });
-
-    it("should return false if dataset is not archivable and not retrievable", () => {
-      const dataset = createMock<DatasetClass>({});
-      dataset.datasetlifecycle = {
-        archivable: false,
-        retrievable: false,
-      };
-
-      const retrievable = component.retrievableCondition(dataset);
-
-      expect(retrievable).toEqual(false);
-    });
-
-    it("should return false if dataset is archivable and retrievable", () => {
-      const dataset = createMock<DatasetClass>({});
-      dataset.datasetlifecycle = {
-        archivable: true,
-        retrievable: true,
-      };
-
-      const retrievable = component.retrievableCondition(dataset);
-
-      expect(retrievable).toEqual(false);
-    });
-
-    it("should return true if dataset is retrievable and not archivable", () => {
-      const dataset = createMock<DatasetClass>({});
-      dataset.datasetlifecycle = {
-        archivable: false,
-        retrievable: true,
-      };
-
-      const retrievable = component.retrievableCondition(dataset);
-
-      expect(retrievable).toEqual(true);
     });
   });
 
@@ -404,175 +220,6 @@ describe("DatasetTableComponent", () => {
     });
   });
 
-  describe("#convertSavedColumns() with instrumentName", () => {
-    beforeEach(() => {
-      component.instruments = [
-        {
-          pid: "instrument1",
-          uniqueName: "unique1",
-          name: "Test Instrument 1",
-        },
-        {
-          pid: "instrument2",
-          uniqueName: "unique2",
-          name: "Test Instrument 2",
-        },
-        { pid: "instrument3", uniqueName: "unique3", name: "" },
-      ] as any[];
-
-      component.instrumentMap = new Map(
-        component.instruments.map((instrument) => [instrument.pid, instrument]),
-      );
-    });
-
-    it("should render instrument name when instrument is found", () => {
-      const columns = [
-        {
-          name: "instrumentName",
-          order: 0,
-          enabled: true,
-          width: 200,
-          type: "standard" as const,
-        },
-      ];
-
-      const convertedColumns = component.convertSavedColumns(columns);
-      const instrumentColumn = convertedColumns[0];
-
-      const mockRow = { instrumentId: "instrument1" };
-      const result = instrumentColumn.customRender(instrumentColumn, mockRow);
-
-      expect(result).toBe("Test Instrument 1");
-    });
-
-    it("should render instrumentId when instrument is not found", () => {
-      const columns = [
-        {
-          name: "instrumentName",
-          order: 0,
-          enabled: true,
-          width: 200,
-          type: "standard" as const,
-        },
-      ];
-
-      const convertedColumns = component.convertSavedColumns(columns);
-      const instrumentColumn = convertedColumns[0];
-
-      const mockRow = { instrumentId: "nonexistent" };
-      const result = instrumentColumn.customRender(instrumentColumn, mockRow);
-
-      expect(result).toBe("nonexistent");
-    });
-
-    it("should render '-' when instrumentId is not present", () => {
-      const columns = [
-        {
-          name: "instrumentName",
-          order: 0,
-          enabled: true,
-          width: 200,
-          type: "standard" as const,
-        },
-      ];
-
-      const convertedColumns = component.convertSavedColumns(columns);
-      const instrumentColumn = convertedColumns[0];
-
-      const mockRow = {};
-      const result = instrumentColumn.customRender(instrumentColumn, mockRow);
-
-      expect(result).toBe("-");
-    });
-
-    it("should render instrumentId when instrument has empty name", () => {
-      const columns = [
-        {
-          name: "instrumentName",
-          order: 0,
-          enabled: true,
-          width: 200,
-          type: "standard" as const,
-        },
-      ];
-
-      const convertedColumns = component.convertSavedColumns(columns);
-      const instrumentColumn = convertedColumns[0];
-
-      const mockRow = { instrumentId: "instrument3" };
-      const result = instrumentColumn.customRender(instrumentColumn, mockRow);
-
-      expect(result).toBe("instrument3");
-    });
-
-    it("should export instrument name when instrument is found", () => {
-      const columns = [
-        {
-          name: "instrumentName",
-          order: 0,
-          enabled: true,
-          width: 200,
-          type: "standard" as const,
-        },
-      ];
-
-      const convertedColumns = component.convertSavedColumns(columns);
-      const instrumentColumn = convertedColumns[0];
-
-      const mockRow = { instrumentId: "instrument2" };
-      const result = instrumentColumn.toExport(mockRow, instrumentColumn);
-
-      expect(result).toBe("Test Instrument 2");
-    });
-
-    it("should export instrumentId when instrument is not found", () => {
-      const columns = [
-        {
-          name: "instrumentName",
-          order: 0,
-          enabled: true,
-          width: 200,
-          type: "standard" as const,
-        },
-      ];
-
-      const convertedColumns = component.convertSavedColumns(columns);
-      const instrumentColumn = convertedColumns[0];
-
-      const mockRow = { instrumentId: "unknown-instrument" };
-      const result = instrumentColumn.toExport(mockRow, instrumentColumn);
-
-      expect(result).toBe("unknown-instrument");
-    });
-
-    it("should not affect other column types", () => {
-      const columns = [
-        {
-          name: "datasetName",
-          order: 0,
-          enabled: true,
-          width: 200,
-          type: "standard" as const,
-        },
-        {
-          name: "instrumentName",
-          order: 1,
-          enabled: true,
-          width: 200,
-          type: "standard" as const,
-        },
-      ];
-
-      const convertedColumns = component.convertSavedColumns(columns);
-
-      expect(convertedColumns.length).toBe(2);
-      expect(convertedColumns[0].name).toBe("datasetName");
-      expect(convertedColumns[0].customRender).toBeUndefined();
-      expect(convertedColumns[1].name).toBe("instrumentName");
-      expect(convertedColumns[1].customRender).toBeDefined();
-    });
-  });
-
   describe("instruments subscription with Map optimization", () => {
     it("should update both instruments array and instrumentMap when instruments observable changes", () => {
       const mockInstruments = [
@@ -628,84 +275,6 @@ describe("DatasetTableComponent", () => {
 
       const notFoundInstrument = component.instrumentMap.get("nonexistent");
       expect(notFoundInstrument).toBeUndefined();
-    });
-  });
-
-  describe("#getInstrumentName() private method", () => {
-    beforeEach(() => {
-      const mockInstruments = [
-        {
-          ...auditFields,
-          pid: "inst1",
-          uniqueName: "unique1",
-          name: "Test Instrument 1",
-        },
-        {
-          ...auditFields,
-          pid: "inst2",
-          uniqueName: "unique2",
-          name: "Test Instrument 2",
-        },
-        { ...auditFields, pid: "inst3", uniqueName: "unique3", name: "" },
-      ];
-
-      component.instrumentMap = new Map(
-        mockInstruments.map((instrument) => [instrument.pid, instrument]),
-      );
-    });
-
-    it("should return instrument name when instrument is found", () => {
-      const mockRow = { instrumentId: "inst1" } as any;
-      const result = component["getInstrumentName"](mockRow);
-      expect(result).toBe("Test Instrument 1");
-    });
-
-    it("should return instrumentId when instrument is not found", () => {
-      const mockRow = { instrumentId: "nonexistent" } as any;
-      const result = component["getInstrumentName"](mockRow);
-      expect(result).toBe("nonexistent");
-    });
-
-    it("should return '-' when instrumentId is not present", () => {
-      const mockRow = {} as any;
-      const result = component["getInstrumentName"](mockRow);
-      expect(result).toBe("-");
-    });
-
-    it("should return instrumentId when instrument has empty name", () => {
-      const mockRow = { instrumentId: "inst3" } as any;
-      const result = component["getInstrumentName"](mockRow);
-      expect(result).toBe("inst3");
-    });
-
-    it("should handle undefined instrumentId gracefully", () => {
-      const mockRow = { instrumentId: undefined } as any;
-      const result = component["getInstrumentName"](mockRow);
-      expect(result).toBe("-");
-    });
-
-    it("should handle null instrumentId gracefully", () => {
-      const mockRow = { instrumentId: null } as any;
-      const result = component["getInstrumentName"](mockRow);
-      expect(result).toBe("-");
-    });
-
-    it("should handle empty string instrumentId gracefully", () => {
-      const mockRow = { instrumentId: "" } as any;
-      const result = component["getInstrumentName"](mockRow);
-      expect(result).toBe("-");
-    });
-
-    it("should return instrument name even when instrumentId is empty but instrument exists", () => {
-      // Add an instrument with empty string pid to test edge case
-      component.instrumentMap.set("", {
-        pid: "",
-        name: "Empty PID Instrument",
-      } as any);
-
-      const mockRow = { instrumentId: "" } as any;
-      const result = component["getInstrumentName"](mockRow);
-      expect(result).toBe("Empty PID Instrument");
     });
   });
 });
