+++ src/app/datasets/dataset-table/dataset-table.component.ts
@@ -68,6 +68,7 @@ import { actionMenu } from "shared/modules/dynamic-material-table/utilizes/defau
 import { TableConfigService } from "shared/services/table-config.service";
 import { selectInstruments } from "state-management/selectors/instruments.selectors";
 import { FormatNumberPipe } from "shared/pipes/format-number.pipe";
+import { DatasetsListService } from "shared/services/datasets-list.service";
 
 export interface SortChangeEvent {
   active: string;
@@ -159,23 +160,9 @@ export class DatasetTableComponent implements OnInit, OnDestroy {
     public appConfigService: AppConfigService,
     private store: Store,
     private route: ActivatedRoute,
-    private jsonHeadPipe: JsonHeadPipe,
-    private datePipe: DatePipe,
-    private fileSize: FileSizePipe,
     private tableConfigService: TableConfigService,
-    private formatNumberPipe: FormatNumberPipe,
+    private datasetsListService: DatasetsListService,
   ) {}
-
-  private getInstrumentName(row: OutputDatasetObsoleteDto): string {
-    const instrument = this.instrumentMap.get(row.instrumentId);
-    if (instrument?.name) {
-      return instrument.name;
-    }
-    if (row.instrumentId != null) {
-      return row.instrumentId === "" ? "-" : row.instrumentId;
-    }
-    return "-";
-  }
 
   getTableSort(): ITableSetting["tableSort"] {
     const { queryParams } = this.route.snapshot;
@@ -295,68 +282,6 @@ export class DatasetTableComponent implements OnInit, OnDestroy {
     });
   }
 
-  // conditional to asses dataset status and assign correct icon ArchViewMode.work_in_progress
-  // TODO: when these concepts stabilise, we should move the definitions to site config
-  wipCondition(dataset: DatasetClass): boolean {
-    if (
-      !dataset.datasetlifecycle.archivable &&
-      !dataset.datasetlifecycle.retrievable &&
-      dataset.datasetlifecycle.archiveStatusMessage !==
-        "scheduleArchiveJobFailed" &&
-      dataset.datasetlifecycle.retrieveStatusMessage !==
-        "scheduleRetrieveJobFailed"
-    ) {
-      return true;
-    }
-    return false;
-  }
-
-  systemErrorCondition(dataset: DatasetClass): boolean {
-    if (
-      (dataset.datasetlifecycle.retrievable &&
-        dataset.datasetlifecycle.archivable) ||
-      dataset.datasetlifecycle.archiveStatusMessage ===
-        "scheduleArchiveJobFailed" ||
-      dataset.datasetlifecycle.retrieveStatusMessage ===
-        "scheduleRetrieveJobFailed"
-    ) {
-      return true;
-    }
-    return false;
-  }
-
-  userErrorCondition(dataset: DatasetClass): boolean {
-    if (dataset.datasetlifecycle.archiveStatusMessage === "missingFilesError") {
-      return true;
-    }
-    return false;
-  }
-
-  archivableCondition(dataset: DatasetClass): boolean {
-    if (
-      dataset.datasetlifecycle.archivable &&
-      !dataset.datasetlifecycle.retrievable &&
-      dataset.datasetlifecycle.archiveStatusMessage !== "missingFilesError"
-    ) {
-      return true;
-    }
-    return false;
-  }
-
-  retrievableCondition(dataset: DatasetClass): boolean {
-    if (
-      !dataset.datasetlifecycle.archivable &&
-      dataset.datasetlifecycle.retrievable
-    ) {
-      return true;
-    }
-    return false;
-  }
-
-  // isInBatch(dataset: DatasetClass): boolean {
-  //   return this.inBatchPids.indexOf(dataset.pid) !== -1;
-  // }
-
   onSelect(event: MatCheckboxChange, dataset: OutputDatasetObsoleteDto): void {
     if (event.checked) {
       this.store.dispatch(selectDatasetAction({ dataset }));
@@ -379,135 +304,6 @@ export class DatasetTableComponent implements OnInit, OnDestroy {
     this.store.dispatch(sortByColumnAction({ column, direction }));
   }
 
-  convertSavedColumns(columns: TableColumn[]): TableField<any>[] {
-    return columns
-      .filter((column) => column.name !== "select")
-      .map((column) => {
-        const convertedColumn: TableField<any> = {
-          name: column.name,
-          header: column.header,
-          index: column.order,
-          display: column.enabled ? "visible" : "hidden",
-          width: column.width,
-          type: column.type,
-          format: column.format,
-          tooltip: column.tooltip,
-        };
-
-        // NOTE: This is how we render the custom columns if new config is used.
-        if (column.type === "custom") {
-          convertedColumn.customRender = (c, row) =>
-            lodashGet(row, column.path || column.name);
-          convertedColumn.toExport = (row) =>
-            lodashGet(row, column.path || column.name);
-        }
-
-        if (column.name === "size") {
-          convertedColumn.customRender = (column, row) =>
-            this.fileSize.transform(row[column.name]);
-          convertedColumn.toExport = (row) =>
-            this.fileSize.transform(row[column.name]);
-        }
-
-        if (column.name === "creationTime") {
-          convertedColumn.customRender = (column, row) =>
-            this.datePipe.transform(row[column.name]);
-          convertedColumn.toExport = (row) =>
-            this.datePipe.transform(row[column.name]);
-        }
-
-        if (
-          column.name === "metadata" ||
-          column.name === "scientificMetadata"
-        ) {
-          convertedColumn.customRender = (column, row) => {
-            // NOTE: Maybe here we should use the "scientificMetadata" as field name and not "metadata". This should be changed in the backend config.
-            return this.jsonHeadPipe.transform(row["scientificMetadata"]);
-          };
-          convertedColumn.toExport = (row) => {
-            return this.jsonHeadPipe.transform(row["scientificMetadata"]);
-          };
-        }
-
-        if (column.name === "dataStatus") {
-          convertedColumn.renderContentIcon = (column, row) => {
-            if (this.wipCondition(row)) {
-              return "hourglass_empty";
-            } else if (this.archivableCondition(row)) {
-              return "archive";
-            } else if (this.retrievableCondition(row)) {
-              return "archive";
-            } else if (this.systemErrorCondition(row)) {
-              return "error_outline";
-            } else if (this.userErrorCondition(row)) {
-              return "error_outline";
-            }
-
-            return "";
-          };
-
-          convertedColumn.customRender = (column, row) => {
-            if (this.wipCondition(row)) {
-              return "Work in progress";
-            } else if (this.archivableCondition(row)) {
-              return "Archivable";
-            } else if (this.retrievableCondition(row)) {
-              return "Retrievable";
-            } else if (this.systemErrorCondition(row)) {
-              return "System error";
-            } else if (this.userErrorCondition(row)) {
-              return "User error";
-            }
-
-            return "";
-          };
-
-          convertedColumn.toExport = (row) => {
-            if (this.wipCondition(row)) {
-              return "Work in progress";
-            } else if (this.archivableCondition(row)) {
-              return "Archivable";
-            } else if (this.retrievableCondition(row)) {
-              return "Retrievable";
-            } else if (this.systemErrorCondition(row)) {
-              return "System error";
-            } else if (this.userErrorCondition(row)) {
-              return "User error";
-            }
-
-            return "";
-          };
-        }
-
-        if (column.name === "image") {
-          convertedColumn.renderImage = true;
-          convertedColumn.sort = "none";
-        }
-
-        if (column.name === "instrumentName") {
-          convertedColumn.customRender = (column, row) =>
-            this.getInstrumentName(row);
-          convertedColumn.toExport = (row, column) =>
-            this.getInstrumentName(row);
-        }
-
-        if (column.name.startsWith("scientificMetadata.")) {
-          convertedColumn.customRender = (col, row) => {
-            return String(
-              this.formatNumberPipe.transform(lodashGet(row, col.name)),
-            );
-          };
-          convertedColumn.toExport = (row) => {
-            return String(
-              this.formatNumberPipe.transform(lodashGet(row, column.name)),
-            );
-          };
-        }
-
-        return convertedColumn;
-      });
-  }
-
   ngOnInit() {
     this.store.dispatch(fetchInstrumentsAction({ limit: 1000, skip: 0 }));
 
@@ -517,15 +313,6 @@ export class DatasetTableComponent implements OnInit, OnDestroy {
         this.selectionIds = datasets.map((dataset) => {
           return dataset.pid;
         });
-      }),
-    );
-
-    this.subscriptions.push(
-      this.instruments$.subscribe((instruments) => {
-        this.instruments = instruments;
-        this.instrumentMap = new Map(
-          instruments.map((instrument) => [instrument.pid, instrument]),
-        );
       }),
     );
 
@@ -560,10 +347,12 @@ export class DatasetTableComponent implements OnInit, OnDestroy {
                   this.appConfig?.defaultDatasetsListSettings?.columns;
 
                 const userTableConfigColumns =
-                  this.convertSavedColumns(userConfigColumns);
+                  this.datasetsListService.convertSavedDatasetColumns(
+                    userConfigColumns,
+                  );
 
                 this.tableDefaultSettingsConfig.settingList[0].columnSetting =
-                  this.convertSavedColumns(
+                  this.datasetsListService.convertSavedDatasetColumns(
                     defaultConfigColumns as TableColumn[],
                   );
