+++ src/app/proposals/proposal-datasets/proposal-datasets.component.html
@@ -12,13 +12,14 @@
   [pagination]="pagination"
   [showNoData]="showNoData"
   [rowSelectionMode]="rowSelectionMode"
+  [localization]="localization"
   [printConfig]="printConfig"
   [showProgress]="showProgress"
   [showGlobalTextSearch]="showGlobalTextSearch"
   (paginationChange)="onPaginationChange($event)"
   (onRowEvent)="onRowClick($event)"
   (onTableEvent)="onTableEvent($event)"
-  class="mat-elevation-z2"
+  class="mat-elevation-z2 proposal-dataset-table"
   [emptyMessage]="'No datasets available'"
   [emptyIcon]="'folder'"
 >
