+++ src/app/shared/pipes/filesize.pipe.ts
@@ -1,10 +1,11 @@
-import { Pipe, PipeTransform } from "@angular/core";
+import { Injectable, Pipe, PipeTransform } from "@angular/core";
 import { filesize } from "filesize";
 
 @Pipe({
   name: "filesize",
   standalone: false,
 })
+@Injectable()
 export class FileSizePipe implements PipeTransform {
   transform(value: number): any {
     return filesize(value || 0, { round: 0, base: 2 });
