+++ src/app/shared/shared.module.ts
@@ -29,6 +29,7 @@ import { EmptyContentModule } from "./modules/generic-empty-content/empty-conten
 import { JsonformsAccordionRendererService } from "./services/jsonforms-accordion-renderer.service";
 import { TranslateModule } from "@ngx-translate/core";
 import { JsonPreviewDialogModule } from "./modules/json-preview-dialog/json-preview-dialog.module";
+import { DatasetsListService } from "./services/datasets-list.service";
 @NgModule({
   imports: [
     BreadcrumbModule,
@@ -65,6 +66,7 @@ import { JsonPreviewDialogModule } from "./modules/json-preview-dialog/json-prev
     ScicatDataService,
     AttachmentService,
     JsonformsAccordionRendererService,
+    DatasetsListService,
   ],
   exports: [
     BreadcrumbModule,
